add utils and rearrange
